+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Changed
+- GET api/v3/exchangeInfo now supports single or multi-symbol query
+
 ## 1.4.0 - 2021-07-30
 
 ### Added
@@ -20,12 +23,12 @@
     
 ## 1.2.0 - 2021-07-12
 
-### Change
+### Changed
 - Remove default value in the parameters
 
 ## 1.1.1 - 2021-06-24
 
-### Change
+### Changed
 - Upgrade the dependency packages
 
 ## 1.1.0 - 2021-06-23